Add common device ufunc mechanism for all non-CPU devices.
